feat: decouple auth form from modal
